feat: the metrics and sales-per-variant metircs have been merged and other requested changes
