+++ packages/vendure-plugin-sales-per-variant/src/SalesPerVariantPlugin.ts
@@ -4,23 +4,23 @@ import path from 'path';
 import { schema } from './api/schema.graphql';
 import { MetricsResolver } from './api/metrics.resolver';
 import { MetricsService } from './api/metrics.service';
-import {
-  AverageOrderValueMetric,
-  ConversionRateMetric,
-  MetricCalculation,
-  NrOfOrdersMetric,
-} from './api/strategies';
-import { PLUGIN_INIT_OPTIONS } from './constants';
+// import {
+//   AverageOrderValueMetric,
+//   ConversionRateMetric,
+//   MetricCalculation,
+//   NrOfOrdersMetric,
+// } from './api/strategies';
+// import { PLUGIN_INIT_OPTIONS } from './constants';
 
-export interface SalesPerVariantPluginOptions {
-  metrics: MetricCalculation[];
-  /**
-   * Relations to fetch for orders.
-   * Getting many orders with many relation can be heavy on the DB,
-   * so handle with care
-   */
-  orderRelations?: string[];
-}
+// export interface SalesPerVariantPluginOptions {
+//   metrics: MetricCalculation[];
+//   /**
+//    * Relations to fetch for orders.
+//    * Getting many orders with many relation can be heavy on the DB,
+//    * so handle with care
+//    */
+//   orderRelations?: string[];
+// }
 
 @VendurePlugin({
   imports: [PluginCommonModule],
@@ -28,30 +28,9 @@ export interface SalesPerVariantPluginOptions {
     schema,
     resolvers: [MetricsResolver],
   },
-  providers: [
-    MetricsService,
-    {
-      provide: PLUGIN_INIT_OPTIONS,
-      useFactory: () => SalesPerVariantPlugin.options,
-    },
-  ],
+  providers: [MetricsService],
 })
 export class SalesPerVariantPlugin {
-  static options: SalesPerVariantPluginOptions = {
-    metrics: [
-      new ConversionRateMetric(),
-      new AverageOrderValueMetric(),
-      new NrOfOrdersMetric(),
-    ],
-  };
-
-  static init(
-    options: SalesPerVariantPluginOptions
-  ): typeof SalesPerVariantPlugin {
-    this.options = options;
-    return SalesPerVariantPlugin;
-  }
-
   static ui: AdminUiExtension = {
     extensionPath: path.join(__dirname, 'ui'),
     ngModules: [
