+++ packages/vendure-plugin-sales-per-variant/src/ui/metrics-widget.module.ts
@@ -1,0 +1,9 @@
+import { ProductMetricsWidgetComponent } from './product-metrics-widget';
+import { SharedModule } from '@vendure/admin-ui/core';
+import { NgModule } from '@angular/core';
+import { OrderMetricsWidgetComponent } from './order-metrics-widget';
+@NgModule({
+  imports: [SharedModule],
+  declarations: [ProductMetricsWidgetComponent, OrderMetricsWidgetComponent],
+})
+export class MetricsWidgetModule {}
