+++ packages/vendure-plugin-sales-per-variant/test/metrics.spec.ts
@@ -77,12 +77,10 @@ describe('Metrics', () => {
       GET_METRICS,
       { input: { interval: 'MONTHLY' } }
     );
-    expect(metricSummary.length).toEqual(3);
+    expect(metricSummary.length).toEqual(2);
     const aov = metricSummary.find((m) => m.code === 'aov')!;
-    const cvr = metricSummary.find((m) => m.code === 'cvr')!;
     const nrOfOrders = metricSummary.find((m) => m.code === 'nr-of-orders')!;
     expect(aov.entries.length).toEqual(12);
-    expect(cvr.entries.length).toEqual(12);
     expect(nrOfOrders.entries.length).toEqual(12);
     expect(aov.entries[11].value).toEqual(4921.4);
     expect(nrOfOrders.entries[11].value).toEqual(3);
@@ -94,12 +92,10 @@ describe('Metrics', () => {
       GET_METRICS,
       { input: { interval: 'WEEKLY' } }
     );
-    expect(metricSummary.length).toEqual(3);
+    expect(metricSummary.length).toEqual(2);
     const aov = metricSummary.find((m) => m.code === 'aov')!;
-    const cvr = metricSummary.find((m) => m.code === 'cvr')!;
     const nrOfOrders = metricSummary.find((m) => m.code === 'nr-of-orders')!;
     expect(aov.entries.length).toEqual(26);
-    expect(cvr.entries.length).toEqual(26);
     expect(nrOfOrders.entries.length).toEqual(26);
     expect(aov.entries[25].value).toEqual(4921.4);
     expect(nrOfOrders.entries[25].value).toEqual(3);
@@ -112,12 +108,12 @@ describe('Metrics', () => {
       { input: { interval: 'WEEKLY', variantIds: [3] } }
     );
     // console.log(metricSummary);
-    expect(metricSummary.length).toEqual(3);
+    expect(metricSummary.length).toEqual(2);
     const aov = metricSummary.find((m) => m.code === 'aov')!;
-    const cvr = metricSummary.find((m) => m.code === 'cvr')!;
-    const nrOfOrders = metricSummary.find((m) => m.code === 'nr-of-orders')!;
+    const nrOfOrders = metricSummary.find(
+      (m) => m.code === 'nr-of-times-sold'
+    )!;
     expect(aov.entries.length).toEqual(26);
-    expect(cvr.entries.length).toEqual(26);
     expect(nrOfOrders.entries.length).toEqual(26);
     expect(aov.entries[25].value).toEqual(0);
     expect(nrOfOrders.entries[25].value).toEqual(0);
@@ -129,12 +125,12 @@ describe('Metrics', () => {
       GET_METRICS,
       { input: { interval: 'MONTHLY', variantIds: [3] } }
     );
-    expect(metricSummary.length).toEqual(3);
+    expect(metricSummary.length).toEqual(2);
     const aov = metricSummary.find((m) => m.code === 'aov')!;
-    const cvr = metricSummary.find((m) => m.code === 'cvr')!;
-    const nrOfOrders = metricSummary.find((m) => m.code === 'nr-of-orders')!;
+    const nrOfOrders = metricSummary.find(
+      (m) => m.code === 'nr-of-times-sold'
+    )!;
     expect(aov.entries.length).toEqual(12);
-    expect(cvr.entries.length).toEqual(12);
     expect(nrOfOrders.entries.length).toEqual(12);
     expect(aov.entries[11].value).toEqual(0);
     expect(nrOfOrders.entries[11].value).toEqual(0);
@@ -146,12 +142,12 @@ describe('Metrics', () => {
       GET_METRICS,
       { input: { interval: 'WEEKLY', variantIds: [1] } }
     );
-    expect(metricSummary.length).toEqual(3);
+    expect(metricSummary.length).toEqual(2);
     const aov = metricSummary.find((m) => m.code === 'aov')!;
-    const cvr = metricSummary.find((m) => m.code === 'cvr')!;
-    const nrOfOrders = metricSummary.find((m) => m.code === 'nr-of-orders')!;
+    const nrOfOrders = metricSummary.find(
+      (m) => m.code === 'nr-of-times-sold'
+    )!;
     expect(aov.entries.length).toEqual(26);
-    expect(cvr.entries.length).toEqual(26);
     expect(nrOfOrders.entries.length).toEqual(26);
     expect(aov.entries[25].value).toEqual(4921.4);
     expect(nrOfOrders.entries[25].value).toEqual(3);
@@ -163,12 +159,12 @@ describe('Metrics', () => {
       GET_METRICS,
       { input: { interval: 'MONTHLY', variantIds: [1] } }
     );
-    expect(metricSummary.length).toEqual(3);
+    expect(metricSummary.length).toEqual(2);
     const aov = metricSummary.find((m) => m.code === 'aov')!;
-    const cvr = metricSummary.find((m) => m.code === 'cvr')!;
-    const nrOfOrders = metricSummary.find((m) => m.code === 'nr-of-orders')!;
+    const nrOfOrders = metricSummary.find(
+      (m) => m.code === 'nr-of-times-sold'
+    )!;
     expect(aov.entries.length).toEqual(12);
-    expect(cvr.entries.length).toEqual(12);
     expect(nrOfOrders.entries.length).toEqual(12);
     expect(aov.entries[11].value).toEqual(4921.4);
     expect(nrOfOrders.entries[11].value).toEqual(3);
